+++ __init__.py
@@ -23,7 +23,6 @@ if "bpy" in locals():
     imp.reload(preferences)
     imp.reload(progress_bar)
     imp.reload(properties)
-    imp.reload(properties_comfy)
     imp.reload(task_queue)
     imp.reload(ui_panels)
     imp.reload(ui_preset_styles)
@@ -32,6 +31,8 @@ if "bpy" in locals():
     imp.reload(automatic1111_api)
     imp.reload(stability_api)
     imp.reload(stablehorde_api)
+    imp.reload(properties_comfy)
+    imp.reload(ui_panels_comfyui)
 else:
     from . import (
         addon_updater_ops,
@@ -42,13 +43,14 @@ else:
         preferences,
         progress_bar,
         properties,
-        properties_comfy,
         task_queue,
         utils,
+        properties_comfy,
     )
     from .ui import (
         ui_panels,
         ui_preset_styles,
+        ui_panels_comfyui
     )
     from .sd_backends import (
         comfyui_api,
@@ -68,10 +70,11 @@ def register():
     preferences.register()
     progress_bar.register()
     properties.register()
-    properties_comfy.register()
     task_queue.register()
     ui_panels.register()
     ui_preset_styles.register()
+    properties_comfy.register()
+    ui_panels_comfyui.register()
 
 
 def unregister():
@@ -81,11 +84,12 @@ def unregister():
     operators.unregister()
     preferences.unregister()
     progress_bar.unregister()
-    properties_comfy.unregister()
     properties.unregister()
     task_queue.unregister()
     ui_panels.unregister()
     ui_preset_styles.unregister()
+    properties_comfy.unregister()
+    ui_panels_comfyui.unregister()
 
 
 if __name__ == "__main__":
